ARM: 6528/1: Use CTR for the I-cache line size on ARMv7
The current implementation of the v7_coherent_*_range function assumes
that the D and I cache lines have the same size, which is incorrect
architecturally. This patch adds the icache_line_size macro which reads
the CTR register. The main loop in v7_coherent_*_range is split in two
independent loops or the D and I caches. This also has the performance
advantage that the DSB is moved outside the main loop.

Change-Id: Ibba459ab2b9f0310cdbe02ce9adaedd396bb2175
CRs-fixed: 294396
Reported-by: Kevin Sapp <ksapp@quicinc.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
[smuckle@codeaurora.org: added change to Sc45nm workaround]
Signed-off-by: Steve Muckle <smuckle@codeaurora.org>
Signed-off-by: Jin Hong <jinh@codeaurora.org>